Projects/fast/requirements.txt
aiofiles==22.1.0
anyio==3.6.2
APScheduler==3.9.1.post1
cachetools==5.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.1.3
cryptography==3.4.8
dnspython==2.2.1
ecs-logging==2.0.2
elastic-apm==6.18.0
email-to==0.1.0
email-validator==1.3.0
et-xmlfile==1.1.0
fastapi==0.88.0
fastapi_amis_admin==0.4.2
fastapi_scheduler==0.0.11
future==0.18.2
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.2
importlib-metadata==5.1.0
isodate==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
mime==0.1.0
msal==1.15.0
msal-extensions==0.3.0
msrest==0.6.21
numpy
oauthlib==3.1.1
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
openpyxl==3.0.10
orjson==3.8.3
pandas==1.5.2
portalocker==1.7.1
premailer==3.10.0
protobuf==4.21.11
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.10.2
PyJWT==2.2.0
pymongo==3.12.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.9
secure-smtplib==0.1.1
sharepy==2.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlalchemy_database==0.1.0
sqlmodel==0.0.8
starlette==0.22.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.6
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.15.0
zipp==3.11.0
tox
pytest
coverage